+++ plugin/evm/service.go
@@ -18,6 +18,7 @@ import (
 	"github.com/ava-labs/coreth/params"
 	"github.com/ava-labs/coreth/plugin/evm/atomic"
 	"github.com/ava-labs/coreth/plugin/evm/client"
+	"github.com/ava-labs/coreth/utils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/log"
@@ -96,7 +97,7 @@ func (service *AvaxAPI) Version(r *http.Request, _ *struct{}, reply *VersionRepl
 func (service *AvaxAPI) ExportKey(r *http.Request, args *client.ExportKeyArgs, reply *client.ExportKeyReply) error {
 	log.Info("EVM: ExportKey called")
 
-	address, err := atomic.ParseEthAddress(args.Address)
+	address, err := utils.ParseEthAddress(args.Address)
 	if err != nil {
 		return fmt.Errorf("couldn't parse %s to address: %s", args.Address, err)
 	}
@@ -127,7 +128,7 @@ func (service *AvaxAPI) ImportKey(r *http.Request, args *client.ImportKeyArgs, r
 		return errMissingPrivateKey
 	}
 
-	reply.Address = atomic.GetEthAddress(args.PrivateKey).Hex()
+	reply.Address = utils.GetEthAddress(args.PrivateKey).Hex()
 
 	service.vm.ctx.Lock.Lock()
 	defer service.vm.ctx.Lock.Unlock()
