+++ plugin/evm/user.go
@@ -10,6 +10,7 @@ import (
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
 	"github.com/ava-labs/coreth/plugin/evm/atomic"
+	"github.com/ava-labs/coreth/utils"
 	"github.com/ethereum/go-ethereum/common"
 )
 
@@ -70,7 +71,7 @@ func (u *user) putAddress(privKey *secp256k1.PrivateKey) error {
 		return errKeyNil
 	}
 
-	address := atomic.GetEthAddress(privKey) // address the privKey controls
+	address := utils.GetEthAddress(privKey) // address the privKey controls
 	controlsAddress, err := u.controlsAddress(address)
 	if err != nil {
 		return err
